+++ app/src/main/java/com/pnj/saku_planner/kakeibo/data/remote/dto/entity/AccountDto.kt
@@ -18,7 +18,7 @@ data class AccountDto(
     val name: String,
 
     @SerializedName("balance")
-    val balance: Double,
+    val balance: Long,
 
     @SerializedName("description")
     val description: String?,
