+++ app/src/main/java/com/pnj/saku_planner/kakeibo/presentation/models/TransactionUi.kt
@@ -14,5 +14,5 @@ data class TransactionUi(
     val category: String? = null,
     val type: TransactionType,
     val kakeibo: KakeiboCategoryType? = null,
-    val amount: Double
+    val amount: Long
 )
